// frontend/src/adapter/api/models/trip.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-fwe-ss-24
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface Trip
 */
export interface Trip {

    /**
     * @type {any}
     * @memberof Trip
     */
    id: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    name: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    description?: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    startDate: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    endDate: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    participants?: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    destinations?: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    image?: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    createdAt: any;

    /**
     * @type {any}
     * @memberof Trip
     */
    updatedAt: any;
}